requirements.txt
# Core dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0
pandas>=2.0.0

# PDF processing
pymupdf>=1.23.0
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0
python-Levenshtein>=0.20.0
